requirements.txt
fastapi 
uvicorn 
twilio 
pydantic 
python-dotenv